refactor: rename schemas to reply-schemas
